OBG:630 Extra filter removed
