+++ opba-api-security/src/main/java/de/adorsys/opba/api/security/internal/filter/RequestSignatureValidationFilter.java
@@ -3,11 +3,14 @@ package de.adorsys.opba.api.security.internal.filter;
 
 import de.adorsys.opba.api.security.external.domain.DataToSign;
 import de.adorsys.opba.api.security.external.domain.HttpHeaders;
-import de.adorsys.opba.api.security.internal.service.RequestVerifyingService;
 import de.adorsys.opba.api.security.external.domain.OperationType;
+import de.adorsys.opba.api.security.internal.config.OperationTypeProperties;
+import de.adorsys.opba.api.security.internal.service.RequestVerifyingService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.core.annotation.Order;
+import org.springframework.util.AntPathMatcher;
+import org.springframework.web.filter.OncePerRequestFilter;
+import org.springframework.web.util.UrlPathHelper;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
@@ -20,13 +23,17 @@ import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 
 @Slf4j
-@Order(1)
 @RequiredArgsConstructor
-public class RequestSignatureValidationFilter extends AbstractSecurityFilter {
+public class RequestSignatureValidationFilter extends OncePerRequestFilter {
+    public static final String OPBA_BANKING_PATH = "/v1/banking/**";
 
     private final RequestVerifyingService requestVerifyingService;
     private final Duration requestTimeLimit;
     private final ConcurrentHashMap<String, String> consumerKeysMap;
+    private final OperationTypeProperties properties;
+
+    private final AntPathMatcher matcher = new AntPathMatcher();
+    private final UrlPathHelper pathHelper = new UrlPathHelper();
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
@@ -35,9 +42,15 @@ public class RequestSignatureValidationFilter extends AbstractSecurityFilter {
         String requestTimeStamp = request.getHeader(HttpHeaders.X_TIMESTAMP_UTC);
         String operationType = request.getHeader(HttpHeaders.X_OPERATION_TYPE);
         String xRequestSignature = request.getHeader(HttpHeaders.X_REQUEST_SIGNATURE);
+        String requestPath = request.getRequestURI();
+        String expectedPath = properties.getAllowedPath().get(operationType);
+
+        if (isNotAllowedOperation(requestPath, expectedPath)) {
+            log.error("Request operation type is not allowed");
+            return;
+        }
 
         Instant instant = Instant.parse(requestTimeStamp);
-
         String fintechApiKey = consumerKeysMap.get(fintechId);
 
         if (fintechApiKey == null) {
@@ -47,7 +60,6 @@ public class RequestSignatureValidationFilter extends AbstractSecurityFilter {
         }
 
         DataToSign dataToSign = new DataToSign(UUID.fromString(xRequestId), instant, OperationType.valueOf(operationType));
-
         boolean verificationResult = requestVerifyingService.verify(xRequestSignature, fintechApiKey, dataToSign);
 
         if (!verificationResult) {
@@ -65,9 +77,18 @@ public class RequestSignatureValidationFilter extends AbstractSecurityFilter {
         filterChain.doFilter(request, response);
     }
 
+    @Override
+    protected boolean shouldNotFilter(HttpServletRequest request) {
+        return !matcher.match(OPBA_BANKING_PATH, pathHelper.getPathWithinApplication(request));
+    }
+
     private boolean isRequestExpired(Instant operationTime) {
         Instant now = Instant.now();
         return now.plus(requestTimeLimit).isBefore(operationTime)
                        || now.minus(requestTimeLimit).isAfter(operationTime);
+    }
+
+    private boolean isNotAllowedOperation(String requestURI, String expectedPath) {
+        return expectedPath == null || !requestURI.startsWith(expectedPath);
     }
 }
